// src/client/models/TrustMarkType.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { InternalID } from './InternalID';
/**
 * A trust mark type definition.
 */
export type TrustMarkType = {
    /**
     * Internal identifier for this TrustMarkType.
     */
    id: InternalID;
    /**
     * The trust mark type identifier.
     */
    trust_mark_type: string;
    /**
     * Optional human-readable description for this TrustMarkType.
     */
    description?: string;
};